requirements.txt
# Assignment 2 Requirements
# Core deep learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image processing
Pillow>=9.0.0

# Utilities
tqdm>=4.64.0
matplotlib>=3.5.0